requirements.txt
torch
torchvision
ftfy
regex
tqdm
faiss-cpu
git+https://github.com/openai/CLIP.git
pillow
